feat: refactor with more validations

+++ server/src/controllers/finances/delete-finance/delete-finance.ts
@@ -1,3 +1,4 @@
+import prismaClient from '../../../database/prismaClient';
 import { IFinance } from '../../../models/Finance';
 import { IController, HttpRequest, HttpResponse } from '../../protocols';
 import { IDeleteFinancecRepository } from './protocols';
@@ -17,6 +18,20 @@ export class DeleteFinanceController implements IController {
           body: 'id_required',
         };
       }
+
+      const checkIfIdExists = await prismaClient.finance.findFirst({
+        where: {
+          id: Number(id),
+        },
+      });
+
+      if (!checkIfIdExists) {
+        return {
+          statusCode: 400,
+          body: 'finance_not_found',
+        };
+      }
+
       const category = await this.deleteFinanceRepository.deleteFinance(
         Number(id),
       );
@@ -27,9 +42,6 @@ export class DeleteFinanceController implements IController {
       };
     } catch (err) {
       console.error(err);
-      console.log({
-        err,
-      });
       return {
         statusCode: 500,
         body: 'internal_server_error',
